refactor(web): 增加头像的背景颜色处理
抽象获取颜色的方法
