+++ src/shared/utils/file-helper.ts
@@ -14,6 +14,11 @@ export const getOriginalImage = (thumbnailImageUrl: string): string => {
  * @param path 地址路径
  */
 export const getAbsolutePath = (path: string) => {
+  if (typeof path === 'string' && path.startsWith('/src/')) {
+    // 不处理client的内部路径的头像
+    return path;
+  }
+
   return config.file.getAbsolutePath(path);
 };
