+++ src/shared/utils/string-helper.ts
@@ -1,5 +1,7 @@
 import _isString from 'lodash/isString';
 import urlRegex from 'url-regex';
+import config from '@shared/project.config';
+import str2int from 'str2int';
 
 /**
  * 判断字符串是否是一个blobUrl
@@ -22,4 +24,14 @@ export const getUrls = (str: string): string[] => {
  */
 export function is(it: string) {
   return !!it && it !== '0' && it !== 'false';
+}
+
+export function getAvatarColorHex(name: string) {
+  if (!name) {
+    return '#ffffff'; // 如果获取不到名字，则返回白色
+  }
+
+  const color = config.defaultImg.color;
+  const id = str2int(name);
+  return color[id % color.length];
 }
